updating burger listings

+++ src/services/BurgersService.js
@@ -1,4 +1,5 @@
 import { dbContext } from "../db/DbContext.js"
+import { NotFound } from "../utils/Errors.js"
 
 class BurgersService {
   async getAllBurgers() {
@@ -11,8 +12,24 @@ class BurgersService {
     return newBurger
   }
 
+  async updateBurger(burgerId, updatedBurger) {
+    const burger = await dbContext.Burgers.findById(burgerId)
+    if (burger == null) {
+      throw new NotFound("Could not find a burger with id" + burgerId);
+
+    }
+    burger.name = updatedBurger.name
+    burger.price = updatedBurger.price
+    await burger.save()
+    return burger
+  }
+
   async removeBurger(burgerId) {
     const burgerToDelete = await dbContext.Burgers.findById(burgerId)
+    if (burgerToDelete == null) {
+      throw new NotFound("Could not find a burger with id" + burgerId);
+
+    }
     await burgerToDelete.deleteOne()
     return burgerToDelete
   }
